requirements.txt
streamlit
numpy
opencv-python==4.10.0.84
pyzbar
easyocr
setuptools
wheel
Pillow
gspread
oauth2client
python-dotenv
pytz
